Configura a visualização de Currency dentro do fragment_home.xml
